// src/server/model/iStockLocation.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface IStockLocation { 
    typeName?: string;
    mobile?: string;
    email?: string;
    stockLocationtypeId?: number;
    stocklocationtypeName?: string;
    costCenterName?: string;
    phone?: string;
    fax?: string;
    pincode?: string;
    address1?: string;
    cityId?: number;
    address3?: string;
    cityName?: string;
    contactPerson?: string;
    address2?: string;
    storeName?: string;
    storeId?: number;
    costCenterId?: number;
    typeId?: number;
    name?: string;
    description?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    productUserName?: string;
    productUserId?: number;
    logInfo?: string;
    importedId?: number;
    creationdate?: Date;
    id?: number;
    jacksontype?: string;
}