// src/server/model/iTaxLine.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ITaxLine { 
    value?: number;
    amount?: number;
    taxName?: string;
    taxOnTaxLines?: Array<ITaxLine>;
    primaryTaxLine?: boolean;
    fixed?: boolean;
    tax?: number;
    taxOnTaxAmount?: number;
    id?: number;
    jacksontype?: string;
}