// src/server/model/pItemMaster.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PItemMaster { 
    id?: number;
    name?: string;
    productCode?: string;
    groupId?: number;
    groupName?: string;
    unitId?: number;
    unitName?: string;
    mrp?: number;
    sellingPrice?: number;
    taxClassId?: number;
    taxClassName?: string;
    hsnCode?: string;
    service?: boolean;
    wholeSalePrice?: number;
    landingPrice?: number;
    rateCode?: string;
    purchasePrice?: number;
    dealerPrice?: number;
    manufacturerName?: string;
    description?: string;
    openingStock?: number;
    openingRate?: number;
}